+++ src/features/card-detail/ui/cardDetailHeader.tsx
@@ -11,7 +11,7 @@ import Appbar from '@/shared/ui/appbar';
 import { BottomModal } from '@/shared/ui/bottomModal/bottomModal';
 import { BottomMenuItem } from '@/shared/ui/bottomModal/bottomModalItem';
 import BottomModalTitle from '@/shared/ui/bottomModal/bottomModalTitle';
-import { ModeInput } from '@/shared/ui/bottomModal/modeInput';
+import { MemoInput } from '@/shared/ui/bottomModal/memoInput';
 import Tag from '@/shared/ui/tag/tag';
 
 import JOB_CONFIG, { JobType } from '../config/jobs-config';
@@ -120,14 +120,14 @@ const CardDetailHeader = () => {
         </div>
       </div>
       {!mode ? (
-        <BottomModal isModalOpen={isModalOpen} closeModal={closeModal}>
+        <BottomModal isModalOpen={isModalOpen} closeModal={closeModal} mode={mode}>
           <BottomMenuItem onClick={handleMode}>한줄 메모</BottomMenuItem>
           <BottomMenuItem>삭제하기</BottomMenuItem>
         </BottomModal>
       ) : (
-        <BottomModal isModalOpen={isModalOpen} closeModal={handleCancelMode}>
+        <BottomModal isModalOpen={isModalOpen} closeModal={handleCancelMode} mode={mode}>
           <BottomModalTitle>한줄 메모</BottomModalTitle>
-          <ModeInput onClose={closeModal} handleCancelMode={handleCancelMode} />
+          <MemoInput onClose={closeModal} handleCancelMode={handleCancelMode} />
         </BottomModal>
       )}
       <Toaster position="bottom-center" />
